Use Popen.communicate to avoid deadlocks
If the child process generates enough data to fill the pipes it is
possible that no data is send anymore and the child process blocks.
If the starting process is using Popen.wait() it will complete deadlock
and can only be killed by the user.
The python documentation recommends to use Popen.communicate instead
for such applications.